+++ apps/scoutgameadmin/components/repos/ReposDashboard.tsx
@@ -18,7 +18,8 @@ import {
   TextField,
   Box,
   IconButton,
-  TableSortLabel
+  TableSortLabel,
+  Button
 } from '@mui/material';
 import React, { useState, useMemo } from 'react';
 
@@ -27,10 +28,11 @@ import { useSearchRepos } from 'hooks/api/repos';
 import { useDebouncedValue } from 'hooks/useDebouncedValue';
 import type { Repo } from 'lib/repos/getRepos';
 
-import { AddRepoButton } from './AddRepoButton/AddRepoButton';
-import { DeleteRepoButton } from './DeleteRepoButton/DeleteRepoButton';
+import { AddRepoButton } from './components/AddRepoButton/AddRepoButton';
+import { DeleteRepoButton } from './components/DeleteRepoButton/DeleteRepoButton';
+import { HeaderActions } from './components/HeaderActions';
 
-type SortField = 'commits' | 'prs' | 'closedPrs' | 'contributors' | 'owner' | 'createdAt';
+type SortField = 'commits' | 'prs' | 'closedPrs' | 'contributors' | 'owner' | 'createdAt' | 'bonusPartner';
 type SortOrder = 'asc' | 'desc';
 
 export function ReposDashboard({ repos }: { repos: Repo[] }) {
@@ -47,8 +49,10 @@ export function ReposDashboard({ repos }: { repos: Repo[] }) {
       return filteredRepos || [];
     }
     return repos.sort((a, b) => {
-      if (a[sortField] < b[sortField]) return sortOrder === 'asc' ? -1 : 1;
-      if (a[sortField] > b[sortField]) return sortOrder === 'asc' ? 1 : -1;
+      if (!a[sortField]) return sortOrder === 'asc' ? -1 : 1;
+      if (!b[sortField]) return sortOrder === 'asc' ? 1 : -1;
+      if (a[sortField]! < b[sortField]!) return sortOrder === 'asc' ? -1 : 1;
+      if (a[sortField]! > b[sortField]!) return sortOrder === 'asc' ? 1 : -1;
       // sort by name as a secondary sort if the field is the same
       if (a[sortField] === b[sortField]) return a.name.localeCompare(b.name);
       return 0;
@@ -88,14 +92,7 @@ export function ReposDashboard({ repos }: { repos: Repo[] }) {
             }
           }}
         />
-        <Box>
-          <AddRepoButton variant='contained' color='primary' sx={{ mr: 2 }}>
-            Add Repos
-          </AddRepoButton>
-          <ExportButton variant='outlined' filename='github_repos.tsv' src='/api/repos/export'>
-            Export Repos
-          </ExportButton>
-        </Box>
+        <HeaderActions />
       </Stack>
       <TableContainer component={Paper}>
         <Table>
@@ -111,15 +108,6 @@ export function ReposDashboard({ repos }: { repos: Repo[] }) {
                 </TableSortLabel>
               </TableCell>
               <TableCell>Name</TableCell>
-              <TableCell>
-                <TableSortLabel
-                  active={sortField === 'createdAt'}
-                  direction={sortField === 'createdAt' ? sortOrder : 'asc'}
-                  onClick={() => handleSort('createdAt')}
-                >
-                  Imported at
-                </TableSortLabel>
-              </TableCell>
               <TableCell>
                 <TableSortLabel
                   active={sortField === 'commits'}
@@ -156,6 +144,15 @@ export function ReposDashboard({ repos }: { repos: Repo[] }) {
                   Contributors
                 </TableSortLabel>
               </TableCell>
+              <TableCell>
+                <TableSortLabel
+                  active={sortField === 'bonusPartner'}
+                  direction={sortField === 'bonusPartner' ? sortOrder : 'asc'}
+                  onClick={() => handleSort('bonusPartner')}
+                >
+                  Bonus Partner
+                </TableSortLabel>
+              </TableCell>
               <TableCell align='center'>Status</TableCell>
             </TableRow>
           </TableHead>
@@ -172,11 +169,11 @@ export function ReposDashboard({ repos }: { repos: Repo[] }) {
                     {repo.name}
                   </Link>
                 </TableCell>
-                <TableCell>{new Date(repo.createdAt).toLocaleDateString()}</TableCell>
                 <TableCell>{repo.commits}</TableCell>
                 <TableCell>{repo.prs}</TableCell>
                 <TableCell>{repo.closedPrs}</TableCell>
                 <TableCell>{repo.contributors}</TableCell>
+                <TableCell>{repo.bonusPartner ? repo.bonusPartner : ''}</TableCell>
                 <TableCell align='center'>
                   <DeleteRepoButton onDelete={() => mutate()} repoId={repo.id} deletedAt={repo.deletedAt} />
                 </TableCell>
