+++ packages/scoutgame/src/dates.ts
@@ -68,6 +68,11 @@ export function getWeekStartEnd(date: Date) {
   return { start: startOfWeek, end: endOfWeek };
 }
 
+export function getWeekStartEndFormatted(date: Date) {
+  const { start, end } = getWeekStartEnd(date);
+  return `${start.toFormat('MMM, dd')} - ${end.toFormat('MMM, dd')}`;
+}
+
 export function getStartOfSeason(week: ISOWeek) {
   return getDateFromISOWeek(week);
 }
